Unified function library

+++ branches/user/foerg/examples2/hydraulics/hydline_galerkin/system.cc
@@ -37,7 +37,7 @@ System::System(const string &name, bool setvalued) : Group(name) {
   l23->setDirection("[0;0;0]");
 
   ConstrainedNode * n1 = new ConstrainedNode("n1");
-  n1->setpFunction(new ConstantFunction<double(double)>(3e5));
+  n1->setpFunction(new ConstantFunction<double>(3e5));
   addLink(n1);
   n1->addOutFlow(l12);
